feat: patch toString to call the llm
